move BaseImgList into its own module; add tests in ex08
